+++ packages/image_picker/image_picker_ios/example/ios/Runner/AppDelegate.m
@@ -12,7 +12,6 @@
 
 - (BOOL)application:(UIApplication *)application
     didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
-  [GeneratedPluginRegistrant registerWithRegistry:self];
   if (@available(iOS 14, *)) {
     // Seed the photo library with at least one image for tests to operate on.
     NSString *photoAddedKey = @"PhotoAdded";
@@ -43,6 +42,10 @@
     }
   }
   return [super application:application didFinishLaunchingWithOptions:launchOptions];
+}
+
+- (void)didInitializeImplicitFlutterEngine:(NSObject<FlutterImplicitEngineBridge> *)engineBridge {
+  [GeneratedPluginRegistrant registerWithRegistry:engineBridge.pluginRegistry];
 }
 
 @end