+++ packages/material_ui/example/lib/navigation_bar/navigation_bar.2.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// #region body
 import 'package:material_ui/material_ui.dart';
 
 /// Flutter code sample for [NavigationBar] with nested [Navigator] destinations.
@@ -391,3 +392,4 @@ class _DestinationViewState extends State<DestinationView> {
     );
   }
 }
+// #endregion body