+++ packages/material_ui/example/lib/segmented_button/segmented_button.0.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// #region body
 import 'package:material_ui/material_ui.dart';
 
 /// Flutter code sample for [SegmentedButton].
@@ -118,3 +119,4 @@ class _MultipleChoiceState extends State<MultipleChoice> {
     );
   }
 }
+// #endregion body