+++ packages/material_ui/lib/src/date_picker_theme.dart
@@ -243,16 +243,19 @@ class DatePickerThemeData with Diagnosticable {
   /// value of [dayBackgroundColor] is used to paint the shape decoration of
   /// the day label.
   ///
-  // TODO(framework): Replace the following block with a @dartpad directive
-  // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// <callout-box>
   ///
   /// This sample demonstrates how to customize the day selector shape decoration
   /// using the [dayShape], [todayForegroundColor], [todayBackgroundColor], and
   /// [todayBorder] properties.
   ///
-  /// {@example /example/lib/date_picker/date_picker_theme_day_shape.0.dart}
+  // TODO(framework): Replace the following block with a @dartpad directive
+  // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// {@macro material_ui.dartpad_guide}
   ///
-  // TODO(framework): End of the @dartpad directive.
+  /// {@example /example/lib/date_picker/date_picker_theme_day_shape.0.dart#body}
+  ///
+  /// </callout-box>
   final WidgetStateProperty<OutlinedBorder?>? dayShape;
 
   /// Overrides the default color used to paint the
@@ -262,16 +265,19 @@ class DatePickerThemeData with Diagnosticable {
   ///
   /// This will be used instead of the [TextStyle.color] provided in [dayStyle].
   ///
-  // TODO(framework): Replace the following block with a @dartpad directive
-  // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// <callout-box>
   ///
   /// This sample demonstrates how to customize the day selector shape decoration
   /// using the [dayShape], [todayForegroundColor], [todayBackgroundColor], and
   /// [todayBorder] properties.
   ///
-  /// {@example /example/lib/date_picker/date_picker_theme_day_shape.0.dart}
+  // TODO(framework): Replace the following block with a @dartpad directive
+  // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// {@macro material_ui.dartpad_guide}
   ///
-  // TODO(framework): End of the @dartpad directive.
+  /// {@example /example/lib/date_picker/date_picker_theme_day_shape.0.dart#body}
+  ///
+  /// </callout-box>
   final WidgetStateProperty<Color?>? todayForegroundColor;
 
   /// Overrides the default color used to paint the background of the
@@ -287,16 +293,19 @@ class DatePickerThemeData with Diagnosticable {
   /// is used as specified. To omit the border entirely,
   /// set [todayBorder] to [BorderSide.none].
   ///
-  // TODO(framework): Replace the following block with a @dartpad directive
-  // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// <callout-box>
   ///
   /// This sample demonstrates how to customize the day selector shape decoration
   /// using the [dayShape], [todayForegroundColor], [todayBackgroundColor], and
   /// [todayBorder] properties.
   ///
-  /// {@example /example/lib/date_picker/date_picker_theme_day_shape.0.dart}
+  // TODO(framework): Replace the following block with a @dartpad directive
+  // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// {@macro material_ui.dartpad_guide}
   ///
-  // TODO(framework): End of the @dartpad directive.
+  /// {@example /example/lib/date_picker/date_picker_theme_day_shape.0.dart#body}
+  ///
+  /// </callout-box>
   final BorderSide? todayBorder;
 
   /// Overrides the default text style used to paint each of the year