+++ packages/material_ui/lib/src/reorderable_list.dart
@@ -26,13 +26,16 @@ import 'theme.dart';
 /// The [onReorderItem] parameter will be called when a child
 /// widget is dragged to a new position.
 ///
+/// <callout-box>
+///
+///
 // TODO(framework): Replace the following block with a @dartpad directive
 // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+/// {@macro material_ui.dartpad_guide}
 ///
+/// {@example /example/lib/reorderable_list/reorderable_list_view.0.dart#body}
 ///
-/// {@example /example/lib/reorderable_list/reorderable_list_view.0.dart}
-///
-// TODO(framework): End of the @dartpad directive.
+/// </callout-box>
 ///
 /// By default, on [TargetPlatformVariant.desktop] platforms each item will
 /// have a drag handle added on top of it that will allow the user to grab it
@@ -46,8 +49,7 @@ import 'theme.dart';
 /// This example demonstrates using the [ReorderableListView.proxyDecorator] callback
 /// to customize the appearance of a list item while it's being dragged.
 ///
-// TODO(framework): Replace the following block with a @dartpad directive
-// when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+/// <callout-box>
 ///
 /// While a drag is underway, the widget returned by the [ReorderableListView.proxyDecorator]
 /// callback serves as a "proxy" (a substitute) for the item in the list. The proxy is
@@ -55,23 +57,30 @@ import 'theme.dart';
 /// callback in this example is similar to the default one except that it changes the
 /// proxy item's background color.
 ///
-/// {@example /example/lib/reorderable_list/reorderable_list_view.1.dart}
+// TODO(framework): Replace the following block with a @dartpad directive
+// when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+/// {@macro material_ui.dartpad_guide}
 ///
-// TODO(framework): End of the @dartpad directive.
+/// {@example /example/lib/reorderable_list/reorderable_list_view.1.dart#body}
+///
+/// </callout-box>
 ///
 /// This example demonstrates using the [ReorderableListView.proxyDecorator] callback to
 /// customize the appearance of a [Card] while it's being dragged.
 ///
-// TODO(framework): Replace the following block with a @dartpad directive
-// when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+/// <callout-box>
 ///
 /// The default [proxyDecorator] wraps the dragged item in a [Material] widget and animates
 /// its elevation. This example demonstrates how to use the [ReorderableListView.proxyDecorator]
 /// callback to update the dragged card elevation without inserted a new [Material] widget.
 ///
-/// {@example /example/lib/reorderable_list/reorderable_list_view.2.dart}
+// TODO(framework): Replace the following block with a @dartpad directive
+// when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+/// {@macro material_ui.dartpad_guide}
 ///
-// TODO(framework): End of the @dartpad directive.
+/// {@example /example/lib/reorderable_list/reorderable_list_view.2.dart#body}
+///
+/// </callout-box>
 class ReorderableListView extends StatefulWidget {
   /// Creates a reorderable list from a pre-built list of widgets.
   ///
@@ -163,13 +172,17 @@ class ReorderableListView extends StatefulWidget {
   /// This example creates a list using the
   /// [ReorderableListView.builder] constructor. Using the [IndexedWidgetBuilder], The
   /// list items are built lazily on demand.
+  ///
+  /// <callout-box>
+  ///
   // TODO(framework): Replace the following block with a @dartpad directive
   // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// {@macro material_ui.dartpad_guide}
   ///
+  /// {@example /example/lib/reorderable_list/reorderable_list_view.reorderable_list_view_builder.0.dart#body}
   ///
-  /// {@example /example/lib/reorderable_list/reorderable_list_view.reorderable_list_view_builder.0.dart}
+  /// </callout-box>
   ///
-  // TODO(framework): End of the @dartpad directive.
   /// See also:
   ///
   ///   * [ReorderableListView], which allows you to build a reorderable
@@ -273,14 +286,17 @@ class ReorderableListView extends StatefulWidget {
   /// The following sample specifies `buildDefaultDragHandles: false`, and
   /// uses a [Card] at the leading edge of each item for the item's drag handle.
   ///
+  /// <callout-box>
+  ///
+  ///
+  ///
   // TODO(framework): Replace the following block with a @dartpad directive
   // when it's supported. https://github.com/dart-lang/dartdoc/issues/4123
+  /// {@macro material_ui.dartpad_guide}
   ///
+  /// {@example /example/lib/reorderable_list/reorderable_list_view.build_default_drag_handles.0.dart#body}
   ///
-  ///
-  /// {@example /example/lib/reorderable_list/reorderable_list_view.build_default_drag_handles.0.dart}
-  ///
-  // TODO(framework): End of the @dartpad directive.
+  /// </callout-box>
   final bool buildDefaultDragHandles;
 
   /// {@macro flutter.widgets.reorderable_list.padding}
@@ -342,7 +358,7 @@ class ReorderableListView extends StatefulWidget {
   /// {@macro flutter.widgets.scrollable.restorationId}
   final String? restorationId;
 
-  /// {@macro flutter.material.Material.clipBehavior}
+  /// {@macro cupertino_ui.Material.clipBehavior}
   ///
   /// Defaults to [Clip.hardEdge].
   final Clip clipBehavior;