+++ packages/material_ui/lib/src/text_selection_toolbar_text_button.dart
@@ -43,14 +43,14 @@ class TextSelectionToolbarTextButton extends StatelessWidget {
   static const double _kMiddlePadding = 9.5;
   static const double _kEndPadding = 14.5;
 
-  /// {@template flutter.material.TextSelectionToolbarTextButton.child}
+  /// {@template material_ui.TextSelectionToolbarTextButton.child}
   /// The child of this button.
   ///
   /// Usually a [Text].
   /// {@endtemplate}
   final Widget child;
 
-  /// {@template flutter.material.TextSelectionToolbarTextButton.onPressed}
+  /// {@template material_ui.TextSelectionToolbarTextButton.onPressed}
   /// Called when this button is pressed.
   /// {@endtemplate}
   final VoidCallback? onPressed;